+++ src/widgets/header/config/const.ts
@@ -1,6 +1,6 @@
 export const navItems = [
   { label: '홈', path: '/' },
   { label: '미션', path: '/talk' },
-  { label: '마음소식', path: '/contents' },
+  { label: '마음소식', path: '/post' },
   { label: '지원사업', path: '/news' },
 ];
